Move schemas into a directory of their own
