+++ pkg/config/schemas/install_schema.go
