+++ pkg/config/schemas/users_schema_test.go
@@ -4,6 +4,7 @@ import (
 	"strings"
 
 	. "github.com/kairos-io/kairos/pkg/config"
+	. "github.com/kairos-io/kairos/pkg/config/schemas"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
